+++ server/src/main/java/org/allaymc/server/network/processor/PacketProcessorHolder.java
@@ -131,6 +131,7 @@ public final class PacketProcessorHolder {
         registerProcessor(ClientState.IN_GAME, new MapInfoRequestPacketProcessor());
         registerProcessor(ClientState.IN_GAME, new ServerboundDiagnosticsPacketProcessor());
         registerProcessor(ClientState.IN_GAME, new RequestAbilityPacketProcessor());
+        registerProcessor(ClientState.IN_GAME, new NPCRequestPacketProcessor());
     }
 
     @SuppressWarnings("unchecked")